refactor: make agent prompts static for LLM caching and extract tool helpers
- Remove previousError retry mechanism from search_events and search_issues
- Make system prompts static (const) to enable LLM provider caching
- Extract createOtelLookupTool to agent-tools/lookup-otel-semantics.ts
- Extract createDatasetAttributesTool to search-events/utils.ts
- Remove duplicate tool creation code (~130 lines)
- Clean up search-issues: remove unused variables and unreachable code
- Remove unnecessary SQL validation from search-issues agent
- Add comprehensive CLAUDE.md documentation for both AI-powered tools
- Update docs with agent-in-tool pattern and static prompt guidance
