+++ packages/mcp-server/src/agent-tools/lookup-otel-semantics.ts
@@ -1,4 +1,7 @@
+import { tool } from "ai";
+import { z } from "zod";
 import type { SentryApiService } from "../api-client";
+import { UserInputError } from "../errors";
 
 // Import all JSON files directly
 import android from "./data/android.json";
@@ -248,4 +251,51 @@ export async function lookupOtelSemantics(
   }
 
   return response;
+}
+
+/**
+ * Create the otel-semantics-lookup tool for AI agents
+ */
+export function createOtelLookupTool(
+  apiService: SentryApiService,
+  organizationSlug: string,
+  projectId?: string,
+) {
+  return tool({
+    description:
+      "Look up OpenTelemetry semantic convention attributes for a specific namespace. OpenTelemetry attributes are universal standards that work across all datasets.",
+    parameters: z.object({
+      namespace: z
+        .string()
+        .describe(
+          "The OpenTelemetry namespace to look up (e.g., 'gen_ai', 'db', 'http', 'mcp')",
+        ),
+      searchTerm: z
+        .string()
+        .optional()
+        .describe("Optional search term to filter attributes"),
+      dataset: z
+        .enum(["spans", "errors", "logs"])
+        .describe(
+          "REQUIRED: Dataset to check attribute availability in. The agent MUST specify this based on their chosen dataset.",
+        ),
+    }),
+    execute: async ({ namespace, searchTerm, dataset }) => {
+      try {
+        return await lookupOtelSemantics(
+          namespace,
+          searchTerm,
+          dataset,
+          apiService,
+          organizationSlug,
+          projectId,
+        );
+      } catch (error) {
+        if (error instanceof UserInputError) {
+          return `Error: ${error.message}`;
+        }
+        throw error;
+      }
+    },
+  });
 }
